+++ Changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Version 1.16.3 - 1.2.2
+ - fix some server-side errors
+ - improve ancient helmet model
+
 ### Version 1.16.3 - 1.2.1
  - added AncientBlaze entity
      - will drop it's helmet on death